requirements.txt
python-telegram-bot[job-queue]==21.6
telethon==1.36.0
anthropic==0.39.0
openai==1.54.3
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
APScheduler==3.10.4
python-dateutil==2.9.0.post0
pytz==2024.2
aiosqlite==0.20.0
httpx==0.27.2
tenacity==9.0.0
